+++ imageServer/serverSetup.py
@@ -120,7 +120,33 @@ def checkNonCanvasCSV(fname):
     # concat name0 and name1 fields into fullName field
     # strip excess whitespace from those fields
     df["studentName"] = df[name0] + ", " + df[name1]
+
     return df
+
+
+def checkLatinNames(df):
+    """Pass the pandas object and check studentNames encode to Latin-1.
+
+    Print out a warning message for any that are not.
+    """
+    # TODO - make this less eurocentric in the future.
+    problems = []
+    for index, row in df.iterrows():
+        try:
+            tmp = row["studentName"].encode("Latin-1")
+        except UnicodeEncodeError:
+            problems.append(
+                'row {}, number {}, name: "{}"'.format(
+                    index, row["id"], row["studentName"]
+                )
+            )
+    if len(problems) > 0:
+        print("WARNING: The following ID/name pairs contain non-Latin characters:")
+        for X in problems:
+            print(X)
+        return False
+    else:
+        return True
 
 
 class SetUp(QWidget):
@@ -217,9 +243,6 @@ class SetUp(QWidget):
                 print("Extracting columns from Canvas data and renaming")
                 df = df[["Student Number", "Student"]]
                 df.columns = ["id", "studentName"]
-                print("Saving to classlist.csv")
-                df.to_csv("../resources/classlist.csv", index=False)
-                return
             else:  # Is not canvas so check we have required headers
                 df = checkNonCanvasCSV(fname)
                 if df is None:
@@ -231,9 +254,21 @@ class SetUp(QWidget):
                     )
                     return
                 df = df[["id", "studentName"]]
-                print("Saving to classlist.csv")
-                df.to_csv("../resources/classlist.csv", index=False)
-                return
+
+            # check characters in names are latin-1 compatible
+            if not checkLatinNames(df):
+                QMessageBox.question(
+                    self,
+                    "Potential classlist problems",
+                    "The classlist you supplied contains non-Latin characters - see console output. "
+                    "You can proceed, but it might cause problems later. "
+                    "Apologies for the eurocentricity.",
+                    buttons=QMessageBox.Ok,
+                )
+
+            print("Saving to classlist.csv")
+            df.to_csv("../resources/classlist.csv", index=False)
+            return
 
 
 if __name__ == "__main__":